ML/requirements.txt
flask
joblib
numpy
pandas
pytest
requests
scikit-learn==1.6.1
python-dotenv